fix: Only execute if Transaction Deletion Record has been used

+++ erpnext/patches/v13_0/delete_orphaned_tables.py
@@ -7,17 +7,21 @@ import frappe
 from frappe.utils import getdate
 
 def execute():
-    child_doctypes = frappe.get_all('DocType', filters = {'istable': 1}, pluck = 'name')
+    if has_deleted_company_transactions():
+        child_doctypes = frappe.get_all('DocType', filters = {'istable': 1}, pluck = 'name')
 
-    for doctype in child_doctypes:
-        docs = frappe.get_all(doctype, fields=['name', 'parent', 'parenttype'])
+        for doctype in child_doctypes:
+            docs = frappe.get_all(doctype, fields=['name', 'parent', 'parenttype'])
 
-        for doc in docs:
-            if not frappe.db.exists(doc['parenttype'], doc['parent']):
-                frappe.db.delete(doctype, {'name': doc['name']})
+            for doc in docs:
+                if not frappe.db.exists(doc['parenttype'], doc['parent']):
+                    frappe.db.delete(doctype, {'name': doc['name']})
 
-            elif check_for_new_doc_with_same_name_as_deleted_parent(doc, doctype):
-                frappe.db.delete(doctype, {'name': doc['name']})
+                elif check_for_new_doc_with_same_name_as_deleted_parent(doc, doctype):
+                    frappe.db.delete(doctype, {'name': doc['name']})
+
+def has_deleted_company_transactions():
+    return frappe.get_all('Transaction Deletion Record')
 
 def check_for_new_doc_with_same_name_as_deleted_parent(doc, doctype):
     """
